implement update sensor data to db

+++ app/models/models.py
@@ -77,7 +77,7 @@ class Sensor(BaseModel):
     __tablename__ = "sensors"
 
     name = Column(String(50), index=True)
-    value = Column(Integer, default=0)
+    value = Column(Float, default=0)
     unit = Column(String(50), nullable=False)
     robot_id = Column(Integer, ForeignKey("robots.id"))
     robot = relationship("Robot", back_populates="sensors")
